making faster

+++ src/config.py
@@ -37,7 +37,7 @@ LLM_MODEL_NAME: str = "Qwen/Qwen3-0.6B"
 # Numerical precision for model weights.
 # torch.float32  → full precision, higher memory (~2× bfloat16)
 # torch.bfloat16 → halves VRAM/RAM with negligible quality loss on modern GPUs
-LLM_DTYPE: torch.dtype = torch.float32
+LLM_DTYPE: torch.dtype = torch.float16
 
 
 MAX_NEW_TOKENS_EXPAND: int = 50
